nino v2.0.2
[Optimisation] More efficient code generation for unmanaged types and arrays

+++ src/Nino.Core/TypeCollector.cs
@@ -2,9 +2,9 @@ namespace Nino.Core
 {
     public abstract class TypeCollector
     {
-        public const int NullTypeId = 0;
-        public const int StringTypeId = 1;
-        public const int CollectionTypeId = 2;
-        public const int NullableTypeId = 3;
+        public const ushort NullTypeId = 0;
+        public const ushort StringTypeId = 1;
+        public const ushort CollectionTypeId = 2;
+        public const ushort NullableTypeId = 3;
     }
 }
